src/main/java/com/AdeAyme/superMarioBros/model/prize/OneUpMushroom.java
package com.aDeAyme.superMarioBros.model.prize;

import com.aDeAyme.superMarioBros.controller.GameEngine;
import com.aDeAyme.superMarioBros.model.hero.Mario;

import java.awt.image.BufferedImage;

public class OneUpMushroom extends BoostItem {

    public OneUpMushroom(double x, double y, BufferedImage style) {
        super(x, y, style);
        setPoint(200);
    }

    @Override
    public void onTouch(Mario mario, GameEngine engine) {
        mario.acquirePoints(getPoint());
        mario.setRemainingLives(mario.getRemainingLives() + 1);
        engine.playOneUp();
    }
}
